[wip] fluxor pattern for contact page.
