--- CoffeeRoastManagement/Client/Store/EditContact/ContactReducers.cs
@@ -1,51 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Fluxor;
-
-namespace CoffeeRoastManagement.Client.Store.EditContact
-{
-    public static class ContactReducers
-    {
-
-        [ReducerMethod(typeof(ContactSubmitAction))]
-        public static ContactState OnSubmit(ContactState state)
-        {
-            return state with
-            {
-                Submitting = true
-            };
-        }
-
-        [ReducerMethod(typeof(ContactSubmitSuccessAction))]
-        public static ContactState OnSubmitSuccess(ContactState state)
-        {
-            return state with
-            {
-                Submitting = false,
-                Submitted = true
-            };
-        }
-
-        [ReducerMethod]
-        public static ContactState OnSubmitFailure(ContactState state, ContactSubmitFailureAction action)
-        {
-            return state with
-            {
-                Submitting = false,
-                ErrorMessage = action.ErrorMessage
-            };
-        }
-
-        [ReducerMethod]
-        public static ContactState OnEditContact(ContactState state, ContactEditAction action) {
-            return state with
-            {
-                Submitted = false,
-                Submitting = false,
-                Contact = action.Contact
-            };
-        }
-    }
-}
